build(tests): -Werror=narrowing en los tests host y fix del estrechamiento que destapa
run-host-tests.sh anade -Werror=narrowing (un estrechamiento en braced-init suele ser una perdida de precision no intencionada). Destapo un long long->u32 en el braced-init de pointer_offset en tile_scroll.hpp; se envuelve cada expresion en static_cast<u32> (mismo valor, intencion explicita). Bateria host completa OK y demo 101 (tile scroll) build/run/analyze OK.

+++ engine/include/eng/graphics/drivers/tile_scroll.hpp
@@ -694,10 +694,12 @@ private:
 			static_cast<u16>((cam_x[1] - 1u) & 0xfff0u),
 		};
 		const u32 pointer_offset[2] = {
-			static_cast<u32>(input.page[0].y) * tile_size * surface_bytes_per_row +
-				static_cast<u32>(cam_y[0]) * surface_bytes_per_row + input.page[0].x * sizeof(u16) + fetch_x[0] / 8u,
-			static_cast<u32>(input.page[1].y) * tile_size * surface_bytes_per_row +
-				static_cast<u32>(cam_y[1]) * surface_bytes_per_row + input.page[1].x * sizeof(u16) + fetch_x[1] / 8u,
+			static_cast<u32>(static_cast<u32>(input.page[0].y) * tile_size * surface_bytes_per_row +
+					 static_cast<u32>(cam_y[0]) * surface_bytes_per_row +
+					 input.page[0].x * sizeof(u16) + fetch_x[0] / 8u),
+			static_cast<u32>(static_cast<u32>(input.page[1].y) * tile_size * surface_bytes_per_row +
+					 static_cast<u32>(cam_y[1]) * surface_bytes_per_row +
+					 input.page[1].x * sizeof(u16) + fetch_x[1] / 8u),
 		};
 
 		m_display.bplcon1 = dual_playfield
